+++ data/src/main/java/com/squirtles/data/datasource/remote/applemusic/model/Attributes.kt
@@ -1,11 +1,11 @@
-package com.example.spotifyapitest.data
+package com.squirtles.data.datasource.remote.applemusic.model
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
 data class Attributes(
-    @SerialName("name") val trackName: String,
+    @SerialName("name") val songName: String,
     val artistName: String,
     val albumName: String? = null,
     val releaseDate: String,
