+++ data/src/main/java/com/squirtles/data/datasource/remote/applemusic/model/MusicVideoResponse.kt
@@ -1,4 +1,4 @@
-package com.example.spotifyapitest.data
+package com.squirtles.data.datasource.remote.applemusic.model
 
 import kotlinx.serialization.Serializable
