+++ data/src/main/java/com/squirtles/data/datasource/remote/applemusic/model/Songs.kt
@@ -1,10 +1,10 @@
-package com.example.spotifyapitest.data
+package com.squirtles.data.datasource.remote.applemusic.model
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Songs (
+data class Songs(
     @SerialName("next") val next: String? = null,
     @SerialName("data") val data: List<Data>,
 )
